requirements.txt
streamlit
pandas
numpy
mysql-connector-python
scikit-learn
python-dotenv
matplotlib
